Absolutely massive overhaul to classes!
I'm so proud of this one! Previously, configuration settings were held in attributes in the Config object. Config.rag_config was a dictionary that had to fit the RagSchema, and I used that for validation. The issue arises when mutating these values during a chat session. Since the RagSchema is not dynamically enforced, there are values exposed to the user that can be incorrectly set and cause unhandled issues. To remedy this, I did a full rework and redid the whole implementation!

classes.py:
- Add classes RagSettings and ChatSettings
- Add ManifestSchema for reading manifest.json and validating an entry.
- All the rag and chat settings implemented with property and setter decorators to enforce the validity of a mutated value at runtime
- Many hours of troubleshooting and tinkering led to this one, 100% worth it!

chat.py:
- Full migration to the new classes
- These are now stored in Chatbot.config and are only accessed through this object.
- I may deprecate use of self.config with `self.settings = self.config.chat_settings` and the analog for rag, but I'll continue to monitor for bugs before I make the variable names the same as the old implementation
- Add Chatbot.rag_model, works consistently and typechecks much better
- Fixes in chat for Ollama LLM using a slightly different API response body (Potential formatting issues when using Ollama as the LLM within chains)
- Lots of other improvements!